// frontend/src/app/core/api/model/vaultEntryResponse.ts
/**
 * Rev-PasswordManager API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface VaultEntryResponse { 
    id?: number;
    title?: string;
    username?: string;
    websiteUrl?: string;
    categoryId?: number;
    categoryName?: string;
    folderId?: number;
    folderName?: string;
    isFavorite?: boolean;
    createdAt?: string;
    updatedAt?: string;
    strengthScore?: number;
    strengthLabel?: string;
}