// frontend/src/app/core/api/model/trashEntryResponse.ts
/**
 * Rev-PasswordManager API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface TrashEntryResponse { 
    id?: number;
    title?: string;
    websiteUrl?: string;
    categoryName?: string;
    folderName?: string;
    deletedAt?: string;
    expiresAt?: string;
    daysRemaining?: number;
}