// frontend/src/app/core/api/model/folderDTO.ts
/**
 * Rev-PasswordManager API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface FolderDTO { 
    id?: number;
    name?: string;
    parentFolderId?: number;
    subfolders?: Array<FolderDTO>;
    createdAt?: string;
    updatedAt?: string;
}