// frontend/src/app/core/api/model/snapshotResponse.ts
/**
 * Rev-PasswordManager API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface SnapshotResponse {
    id?: number;
    password?: string;
    entryTitle?: string;
    changedAt?: string;
}